Fix bug related to ACF get_field()
When submitting new edits on a revision with a pending revision, the
ACF field associated with the new unchanged revision will not get
properly reset to the original value because of a bug somewhere in
get_field() (it grabs the value from the previous revision id, not the
parent post id). Use get_post_meta() instead.